more work on workout api

+++ src/main/java/com/jayr33n/api/MuscleController.java
@@ -1,9 +1,9 @@
 package com.jayr33n.api;
 
-import com.jayr33n.command.MuscleCreateCommand;
+import com.jayr33n.commands.create.MuscleCreateCommand;
 import com.jayr33n.domain.Muscle;
 import com.jayr33n.exception.EntityNotFoundException;
-import com.jayr33n.repository.MuscleRepository;
+import com.jayr33n.repositories.MuscleRepository;
 import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.http.HttpStatus;
@@ -25,21 +25,23 @@ public class MuscleController {
         return repository.findAll(pageable);
     }
 
-    @Post
-    @Status(HttpStatus.CREATED)
-    public void post(@Body @Valid MuscleCreateCommand command) {
-        repository.save(new Muscle(command.getName()));
-    }
-
     @Get("/{id}")
     public Muscle get(Long id) {
         return repository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(id, Muscle.class));
     }
 
+    @Post
+    @Status(HttpStatus.CREATED)
+    public void post(@Body @Valid MuscleCreateCommand command) {
+        repository.save(new Muscle(command.getName()));
+    }
+
     @Delete("/{id}")
     @Status(HttpStatus.NO_CONTENT)
     public void delete(Long id) {
+        if (!repository.existsById(id))
+            throw new EntityNotFoundException(id, Muscle.class);
         repository.deleteById(id);
     }
 }
